Cleanup date formatting
Loggers provide encoder on which signals get encoded. For that reason JRPCLogger provides its own JSONEncoder.

Signal uses a dedicated date logger for constructing description. For that reason it maintains #descriptionDateFormatter. It can be overriden it it's an eyesore. Signal no longer uses that encoder for encoding itself on arbitrary encoders.