Add ability to supress signals by type
Use `enable(loggingTo:, on:)` and `disable(loggingTo:, on:)` on a Signal's class side to allow/disallow logging of that signal on given logger and beacon. Logger and beacon can be nil to match any...

Before `emit`ing - check if the beacons on which we're about to emit a signal, will actually attempt to handle that signal - i.e. it wasn't disabled. This avoid having to construct signals that will definitely not be logged, as that may become unnecessarily costly at runtime