Generalize Signaling protocol for type checking
This allows any invocation of emit() to check willLog. No need to implement that in numerous places.

The normal path to emitting objects (as opposed to creating a signal instance and emitting that) is to check if the object adheres to Signaling - if so, we create appropriate signal and then emit it. We can now determine the type of signal that is going to be and check with `willLog` to see if it's going to be logged and avoid constructing an instance of a `Signal` subclass.

For signals that don't capture a value - like Identity, MachO, Context, etc - we need to actually implement emit() methods that guard via `willLog`.