Delete backup before creating new one
We may have a name collisions when creating a new backup file. When a collision occurs, we append “ (#)” to the filename, incrementing “#” until we find a filename that doesn’t exist.

(1) This fixes this suffixing logic.

(2) Before creating a backup, we first make room for it - that is, we delete a backup file that would become obsolete after creating a new backup. This helps with maintaining indexes in filenames with collision, so we won’t end up with gaps. This approach is also more favorable as we may want to free up resources before consuming more of them - for example, if our strategy is to be within a certain disk usage…